# openapi.yaml
openapi: 3.0.0
info:
  description: Define an API
  title: Restful API
  version: 1.0.0
paths:
  "/opendb/{sql}":
    post:
      tags:
        - Connect Database(Must be connected first)
      summary: 連接資料庫
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/models.DBinformation"
        description: 資料庫資訊
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                type: string
                example: "Successfully Connect Database"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example: 
                "message": "Connect Database failed"
...
  "/{sql}/addvalue/{tablename}":
    post:
      tags:
        - Table CRUD
      summary: 加入數值至資料表
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: tablename
          in: path
          required: true
          schema:
            type: string
          example: "peples"
        - description: 填入數值(colname,value)
          name: value
          in: query
          schema:
            type: array
            example: ["key1,123456"]
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                type: string
                example: 
                  "Successfully Add Value"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
  "/{sql}/getall/{tablename}":
    get:
      tags:
        - Table CRUD
      summary: 取得所有資料
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: tablename
          in: path
          required: true
          schema:
            type: string
          example: peoples
        - description: 挑選欄位
          name: col
          in: query
          schema:
            type: array
            example: ["key1"]
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.object"
              example: 
                欄位名稱: "欄位數值"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
...
  "/{sql}/delete/{tablename}":
    delete:
      tags:
        - Table CRUD
      summary: 刪除資料表數值
      parameters:
        - description: 資料庫引擎
          name: sql
          in: path
          required: true
          schema:
            type: string
            example: 
              "mysql or mssql"
        - description: 資料表名稱
          name: tablename
          in: path
          required: true
          schema:
            type: string
          example: peoples
        - description: 被刪除的欄位條件(colname,value)
          name: where
          in: query
          schema:
            type: array
            example: ["key1,123456"]
            items:
              type: string
          style: form
          explode: true
      responses:
        "200":
          description: Successfully
          content:
            application/json:
              schema:
                type: string
              example: 
                  "Successfully Delete Value"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/models.Error"
              example:
                message: "資料庫發生錯誤"
...
servers:
  - url: http://localhost:8080/v1
components:
  schemas:
    models.DBinformation:
      type: object
      properties:
        database:
          type: string
          example: 資料庫
        host:
          type: string
          example: ip位址
        maxIdle:
          type: integer
          example: 最大連接池
        maxOpen:
          type: integer
          example: 最大可同時開啟數量
        password:
          type: string
          example: 密碼
        port:
          type: string
          example: 端點
        userName:
          type: string
          example: 使用者名稱
    models.Error:
      type: object
      properties:
        message:
          type: string
    models.object:
      type: object